requirements.txt
fastapi
uvicorn
groq
python-dotenv
python-multipart
python-docx
fpdf
PyMuPDF
